refactor: move code for commands to utils classes
