+++ packages/cli/src/commands/deviceprofiles/update.ts
@@ -1,27 +1,12 @@
 import { Errors } from '@oclif/core'
 
-import { DeviceProfile, DeviceProfileRequest } from '@smartthings/core-sdk'
+import { DeviceProfile } from '@smartthings/core-sdk'
 
-import { ActionFunction, APIOrganizationCommand, inputAndOutputItem, TableGenerator } from '@smartthings/cli-lib'
+import { ActionFunction, APIOrganizationCommand, inputAndOutputItem } from '@smartthings/cli-lib'
 
-import { chooseDeviceProfile } from '../deviceprofiles'
-import { DeviceDefinitionRequest } from './view'
+import { buildTableOutputWithoutPreferences, chooseDeviceProfile, cleanupDeviceProfileRequest } from '../../lib/commands/deviceprofiles-util'
+import { DeviceDefinitionRequest } from '../../lib/commands/deviceprofiles/view-util'
 
-
-export function buildTableOutput(tableGenerator: TableGenerator, data: DeviceProfile): string {
-	const table = tableGenerator.newOutputTable()
-	table.push(['Name', data.name])
-	for (const comp of data.components) {
-		table.push([`${comp.id} component`,  comp.capabilities ? comp.capabilities.map(it => it.id).join('\n') : ''])
-	}
-	table.push(['Id', data.id])
-	table.push(['Device Type', data.metadata?.deviceType ?? ''])
-	table.push(['OCF Device Type', data.metadata?.ocfDeviceType ?? ''])
-	table.push(['Manufacturer Name', data.metadata?.mnmn ?? ''])
-	table.push(['Presentation ID', data.metadata?.vid ?? ''])
-	table.push(['Status', data.status])
-	return table.toString()
-}
 
 export default class DeviceProfileUpdateCommand extends APIOrganizationCommand<typeof DeviceProfileUpdateCommand.flags> {
 	static description = 'update a device profile'
@@ -45,25 +30,8 @@ export default class DeviceProfileUpdateCommand extends APIOrganizationCommand<t
 				throw new Errors.CLIError('Input contains "view" property. Use deviceprofiles:view:update instead.')
 			}
 
-			return this.client.deviceProfiles.update(id, cleanupRequest(data))
+			return this.client.deviceProfiles.update(id, cleanupDeviceProfileRequest(data))
 		}
-		await inputAndOutputItem(this, { buildTableOutput: data => buildTableOutput(this.tableGenerator, data) }, executeUpdate)
+		await inputAndOutputItem(this, { buildTableOutput: data => buildTableOutputWithoutPreferences(this.tableGenerator, data) }, executeUpdate)
 	}
-}
-
-// Cleanup is done so that the result of a device profile get can be modified and
-// used in an update operation without having to delete the status, owner, and
-// component name fields, which aren't accepted in the update API call.
-export function cleanupRequest(deviceProfileRequest: Partial<DeviceProfile & { restrictions: unknown }>): DeviceProfileRequest {
-	delete deviceProfileRequest.id
-	delete deviceProfileRequest.status
-	delete deviceProfileRequest.name
-	if (deviceProfileRequest.components) {
-		for (const component of deviceProfileRequest.components) {
-			delete component.label
-		}
-	}
-	delete deviceProfileRequest.restrictions
-
-	return deviceProfileRequest
 }
